fastapi_pt/requirements.txt
fastapi
uvicorn
numpy
pillow
torch
torchvision